+++ R/openmp.R
@@ -30,12 +30,34 @@
 #' specific requirements. As a default, \code{fst} uses a number of threads equal to the number of
 #' logical cores in the system.
 #'
+#' The number of threads can also be set with \code{option(fst_threads = N)}.
+#' NOTE: This option is only read when the package's namespace is first loaded, with commands like
+#' \code{library}, \code{require}, or \code{::}. If you have already used one of these, you
+#' must use \code{threads_fst} to set the number of threads.
+#'
 #' @param nr_of_threads number of threads to use or \code{NULL} to get the current number of threads used in
 #' multithreaded operations.
+#' @param reset_after_fork when \code{fst} is running in a forked process, the usage of OpenMP can
+#' create problems. To prevent these, \code{fst} switches back to single core usage when it detects a fork.
+#' After the fork, the number of threads is reset to it's initial setting. However, on some compilers
+#' (e.g. Intel), switching back to multi-threaded mode can lead to issues. When \code{reset_after_fork}
+#' is set to \code{FALSE}, \code{fst} is left in single-threaded mode after the fork ends. After the fork,
+#' multithreading can be activated again manually by calling \code{threads_fst} with an appropriate value
+#' for \code{nr_of_threads}. The default (\code{reset_after_fork = NULL}) leaves the fork behavior unchanged.
 #'
 #' @return the number of threads (previously) used
 #' @export
-threads_fst <- function(nr_of_threads = NULL) {
+threads_fst <- function(nr_of_threads = NULL, reset_after_fork = NULL) {
+
+  if (!is.null(reset_after_fork)) {
+    if (!is.logical(reset_after_fork) || length(reset_after_fork) != 1 || is.na(reset_after_fork)) {
+      stop("Parameter reset_after_fork should be set to TRUE or FALSE")
+    }
+
+    # change unfork behavior
+    restore_after_fork(reset_after_fork)
+  }
+
   if (is.null(nr_of_threads)) {
     return(getnrofthreads())
   }